+++ fundamentals_screener/chart_generation/clean_web_data.py
